No need to type erase the `Presentation`
The VC itself is already typed erased using the
`_ViewController` base protocol, giving access
to the presented `ContentView` of the specific
`ViewController`.
We would only need this, if the presentation itself
would be allowed to carry a custom `View`. But to
do things like that, we have the `.custom`
presentation mode.