Add a `BindablePublished` property wrapper
An `@Published` like property wrapper, which has a `Binding`
as its `projectedValue`.
This in itself seems to work OK and addresses issue #4.

However, I'd also like to override `@State`, so that this
works:
```swift
class Page: ViewController {
  @State var title = "Hello"

  var view: some View {
    TextField("Title", $title)
  }
}
```

Using a simple:
```swift
extension ViewController {
  public typealias State = BindablePublished
}
```

But this fails, the compiler gets confused in finding the
right `@State` subscripts I presume.